requirements.txt
# ── Dashboard ──
streamlit==1.32.0
pyvis==0.3.2
plotly==5.19.0

# ── Data ──
pandas==2.2.1
numpy==1.26.4
networkx==3.3
matplotlib==3.8.4
seaborn==0.13.2

# ── Database ──
neo4j==5.18.0

# ── Groq LLM ──
groq==0.9.0

# ── LangChain stack (all must match) ──
langchain==0.2.16
langchain-core==0.2.38
langchain-groq==0.1.9
langchain-community==0.2.16
langchain-pinecone==0.1.3
langchain-huggingface==0.0.3

# ── Embeddings & Vector Search ──
sentence-transformers==2.7.0
faiss-cpu==1.8.0

# ── Vector Store (Cloud) ──
pinecone-client==3.2.2

# ── Utils ──
python-dotenv==1.0.1